requirements.txt
python >= 3.6-3.9
numpy
matplotlib
torchvision
tensorflow
pandas

